requirements.txt
chardet==4.0.0
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
numpy==1.21.2
openpyxl==3.0.8
pandas==1.3.3
pdfminer.six==20200517
pdfplumber==0.5.28
Pillow==8.3.2
protobuf==3.17.3
pycryptodome==3.10.1
python-dateutil==2.8.2
pytz==2021.1
reportlab==3.6.1
six==1.16.0
sortedcontainers==2.4.0
typing-extensions==3.10.0.2
waitress==2.0.0
Wand==0.6.7
Werkzeug==2.0.1
zipp==3.5.0
